+++ docker-compose.yml
@@ -73,7 +73,7 @@ services:
         - postgis
         - rabbitmq
       environment:
-        - CONCURRENCY=1
+        - CONCURRENCY=4
         - MAPPROXY_CONCURRENCY=4
       command: sh /var/lib/eventkit/scripts/celery-entrypoint.sh
   celery-beat:
